pyproject.toml
[project]
name = "obsidian-fastmcp"
version = "0.1.0"
description = "A Model Control Protocol (MCP) for Claude Desktop that allows you to create and manage notes in your Obsidian vault"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["obsidian", "claude", "mcp", "notes"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: News/Diary",
]
dependencies = [
    "fastmcp>=0.1.0",
    "genanki>=1.20.0",
    "pydantic>=2.5.2",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0"
]

[project.urls]
Repository = "https://github.com/yourusername/obsidian-fastmcp"
Issues = "https://github.com/yourusername/obsidian-fastmcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
